+++ app/Middleware/RedirectIfAuthenticated.php
@@ -2,6 +2,7 @@
 
 namespace App\Middleware;
 
+use App\Models\User;
 use Niang\Core\Auth;
 use Niang\Core\Http\Request;
 use Niang\Core\Http\Response;
@@ -12,7 +13,7 @@ class RedirectIfAuthenticated implements Middleware
     public function handle(Request $request, \Closure $next): Response
     {
         if (Auth::check()) {
-            return Response::redirect('/');
+            return Response::redirect(User::homePath(Auth::user()));
         }
 
         return $next($request);